+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Warn the user about a potential data leak on mail bounce or forward (#7993)
 - Make the Empty action available for every non-empty folder, not only Trash (#7948)
 - Remove (incorrect) use of Return-Receipt-To header (#8069)
+- Submit various simple dialog forms with the Enter key (#7133)
 - Plugin API: Allow modification of 'error' argument in 'message_send_error' hook (#7914)
 - OAuth: add plugin hooks `oauth_login` and `oauth_refresh_token` for oauth events (#8028, #8040)
 - Debug_logger: Fix the main plugin functionality and documentation (#8041)